+++ src/main/java/nextstep/sessions/repository/RegistrationRepository.java
@@ -7,15 +7,15 @@ import nextstep.sessions.domain.data.registration.Registration;
 
 public interface RegistrationRepository {
 
-    List<Registration> findAllById(int sessionId);
+    List<Registration> findAllById(Long sessionId);
 
     void save(Registration registration);
 
-    Optional<Registration> findById(int registrationId);
+    Optional<Registration> findById(Long registrationId);
 
     void updateSelectionType(Registration registration);
 
     void updateApprovalType(Registration registration);
 
-    void deleteById(int registrationId);
+    void deleteById(Long registrationId);
 }
